# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Northlane Auto Demo API",
    "version": "1.0.0",
    "description": "The complete HTTP contract of the Northlane Auto Insurance demo.\n\nThis is a fictional insurer built for QA automation training. No policy exists, no claim is real, and no money moves.\n\n**Signing in.** `POST /api/auth/login` with `skipMfa: true` and one of the two fixed demo accounts returns a session cookie in one call. Without it, the endpoint opens a two-step verification challenge and answers with `codeDelivery`, telling you whether the code was emailed or is one of the documented fixed codes:\n\n| Account | Password | Fixed code |\n| --- | --- | --- |\n| `customer.demo@testrigor-mail.com` | `CustomerDemo!2026` | `111111` |\n| `agent.demo@testrigor-mail.com` | `AgentDemo!2026` | `222222` |\n\nAn account you create yourself receives an emailed code when a mail provider is configured, and the fallback code `123456` when one is not.\n\n**Driving the workflow.** Every business transition goes through `PATCH /api/demo-state`. The state is a single row shared by every session, so the policyholder and the agent see each other's changes. `DELETE /api/demo-state` restores the seed data.\n\n**Status codes carry meaning.** `403` means the signed-in role may never perform that action. `400` means the request is malformed. `409` means the request is well-formed but the state is not ready for it."
  },
  "servers": [
...
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "login",
        "summary": "Sign in, or open a two-step verification challenge",
        "description": "Signs in a fixed demo account directly when `skipMfa` is true. Otherwise opens a verification challenge and reports how the code was delivered. An email that matches no account is treated as a registration when the password is at least 8 characters.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "examples": {
                "fixedAccountBypass": {
                  "summary": "Fixed demo account, no verification step",
                  "value": { "email": "customer.demo@testrigor-mail.com", "password": "CustomerDemo!2026", "role": "customer", "skipMfa": true }
                },
                "fixedAccountChallenge": {
                  "summary": "Fixed demo account, verification code 111111",
                  "value": { "email": "customer.demo@testrigor-mail.com", "password": "CustomerDemo!2026", "role": "customer" }
                },
                "agentAccount": {
                  "summary": "Claims agent, verification code 222222",
                  "value": { "email": "agent.demo@testrigor-mail.com", "password": "AgentDemo!2026", "role": "agent", "skipMfa": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A session was created, or a verification challenge was opened.",
            "headers": {
              "Set-Cookie": {
                "description": "Present only when `skipMfa` created a session.",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/SessionResponse" },
                    { "$ref": "#/components/schemas/ChallengeResponse" }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "verifyMfa",
        "summary": "Exchange a verification code for a session",
        "description": "Consumes the challenge. A code can be used once, expires after 10 minutes, and locks the challenge after 5 incorrect attempts.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyRequest" },
              "examples": {
                "policyholder": {
                  "summary": "The policyholder's fixed code",
                  "value": { "challengeId": "3f6c9a1e-0d2b-4b6f-9f3c-2a1b7d8e5c40", "code": "111111" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The code matched and a session cookie was set.",
            "headers": { "Set-Cookie": { "schema": { "type": "string" } } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionResponse" } } }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" },
          "410": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth/account": {
      "delete": {
        "tags": ["Authentication"],
        "operationId": "deleteAccount",
        "summary": "Delete the signed-in account",
        "description": "Removes the user, its pending registrations and its verification challenges. The two fixed demo accounts are protected so shared QA credentials always work. The shared workflow state is never affected.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteAccountRequest" },
              "examples": {
                "confirmed": {
                  "summary": "Password plus the literal word DELETE",
                  "value": { "password": "MyDemoPassword1", "confirmation": "DELETE" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The account was deleted and the session cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["ok"],
                  "properties": { "ok": { "type": "boolean", "const": true } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "northlane_session",
        "description": "An HTTP-only signed cookie set by login or verify. A browser sends it automatically."
      }
    },
    "responses": {
      "Error": {
        "description": "The request was refused. The message is written for an integrator.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "examples": ["Only a claim in review can be decided."] }
        }
      },
      "Role": {
        "type": "string",
        "enum": ["customer", "agent"],
        "description": "`customer` is the policyholder; `agent` is the claims agent."
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "name", "role"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" }
        }
      },
      "SessionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user"],
        "properties": { "user": { "$ref": "#/components/schemas/User" } }
      },
      "ChallengeResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["challengeId", "destination", "expiresInSeconds", "codeDelivery"],
        "properties": {
          "challengeId": { "type": "string", "format": "uuid" },
          "destination": { "type": "string", "description": "The masked email the challenge belongs to.", "examples": ["cus•••••••••@testrigor-mail.com"] },
          "expiresInSeconds": { "type": "integer", "examples": [600] },
          "codeDelivery": {
            "type": "string",
            "enum": ["fixed", "email"],
            "description": "`fixed` means the code is one of the documented constants and no email was sent. `email` means a real message was delivered."
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "role": { "$ref": "#/components/schemas/Role" },
          "skipMfa": {
            "type": "boolean",
            "default": false,
            "description": "Fixed demo accounts only. Any other account gets 403."
          }
        }
      },
      "VerifyRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["challengeId", "code"],
        "properties": {
          "challengeId": { "type": "string", "format": "uuid" },
          "code": { "type": "string", "pattern": "^\\d{6}$" }
        }
      },
      "DeleteAccountRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["password", "confirmation"],
        "properties": {
          "password": { "type": "string" },
          "confirmation": { "type": "string", "const": "DELETE" }
        }
      },
...
    }
  },
  "security": [{ "sessionCookie": [] }]
}